Normal changes

+++ database/migrations/2020_06_02_091227_create_student_mnnits_table.php
@@ -18,7 +18,7 @@ class CreateStudentMnnitsTable extends Migration
             $table->string('name');
             $table->string('regno');
             $table->string('email');
-            $table->integer('phoneno');
+            $table->bigInteger('phoneno');
             $table->string('hostel');
             $table->integer('room');
